ai-service/requirements.txt
# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0

# FastAPI and Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0

# Face Recognition & Detection
facenet-pytorch>=2.5.3
insightface>=0.7.3
onnxruntime>=1.16.0

# Metrics & Evaluation
scikit-image>=0.21.0
lpips>=0.1.4

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.1
tqdm>=4.66.0